Rename lcao to LCAO in QMCWaveFunctions.
