+++ src/Numerics/tests/test_full_soa_cartesian_tensor.cpp.in
@@ -12,7 +12,7 @@
 
 #include "catch.hpp"
 #include "Message/Communicate.h"
-#include "QMCWaveFunctions/lcao/SoaCartesianTensor.h"
+#include "QMCWaveFunctions/LCAO/SoaCartesianTensor.h"
 
 #include <stdio.h>
 #include <string>
