+++ src/QMCWaveFunctions/LCAO/MultiFunctorAdapter.h
