+++ src/QMCWaveFunctions/LCAO/NGFunctor.h
