+++ src/QMCWaveFunctions/LCAO/SoaLocalizedBasisSet.h
