+++ src/QMCWaveFunctions/LCAO/SoaSphericalTensor.h
