+++ src/QMCWaveFunctions/tests/test_multiquintic_spline.cpp
@@ -18,11 +18,11 @@
 #include "Message/Communicate.h"
 #include "OhmmsPETE/OhmmsMatrix.h"
 #include "Numerics/OneDimQuinticSpline.h"
-#include "QMCWaveFunctions/lcao/MultiQuinticSpline1D.h"
+#include "QMCWaveFunctions/LCAO/MultiQuinticSpline1D.h"
 
 namespace qmcplusplus
 {
-TEST_CASE("LogGridLight", "[wavefunction][lcao]")
+TEST_CASE("LogGridLight", "[wavefunction][LCAO]")
 {
   LogGridLight<double> grid;
   grid.set(0.1, 1.0, 5);
@@ -43,7 +43,7 @@ TEST_CASE("LogGridLight", "[wavefunction][lcao]")
 }
 
 
-TEST_CASE("MultiQuinticSpline", "[wavefunction][lcao]")
+TEST_CASE("MultiQuinticSpline", "[wavefunction][LCAO]")
 {
   Vector<double> data(5);
   data[0] = 0.0;
